Rewrote class-parameters to be injected as generics, instead of as port-signals.
Added constructors to the AST elements to allow easier custom code construction.

Rewrote the logic for components to support a simple SME based implementation, and then allow an override version to be created with a custom VHDL renderer.
The example implementations are written for 3 variations of block rams, with renders that are based on constructing compiler-inferable VHDL as well as explicit macro instantiation for Xilinx devices.